Move maven dependency get into spec
Moving the mvn calls into the spec file instead of the srpm creation.
This because the srpm is not always created with the same OS (version)
as the ovirt-engine build. For example on COPR the srpm is created on
Fedora, and this causes issues when generating the maven dependencies.
Resulting in missing dependencies when building ovirt-engine on the c9s
chroot.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>