Fix Additional Dependencies on COPR
Fix including ADDITIONAL_DEPENDENCIES on COPR.
It works fine if the srpm and rpm are build on the same
container/directory, as then the file gets created and can be included
in the spec when rpmbuild is started.
But on COPR for example we build the rpm in another chroot than the
srpm, which causes the file to not exist at the start, and results in
the additional dependencies to not be included.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>